refactor: move auth to providers
