+++ studio/src/app/providers/auth/auth.factory.service.ts
@@ -1,6 +1,6 @@
 import {EnvironmentAppConfig} from '../../types/core/environment-config';
 
-import {EnvironmentConfigService} from '../environment/environment-config.service';
+import {EnvironmentConfigService} from '../../services/environment/environment-config.service';
 
 import {AuthService} from './auth.service';
 import {AuthOfflineService} from './auth.offline.service';
